// OLD_REPOP/lib/core/src/lib/dialogs/dialog/dialog-data.interface.ts
/*!
 * @license
 * Copyright © 2005-2025 Hyland Software, Inc. and its affiliates. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { TemplateRef, Type } from '@angular/core';
import { Subject } from 'rxjs';
import { DialogSizes } from './dialog.model';

export interface DialogData {
    title: string;
    description?: string;
    confirmButtonTitle?: string;
    cancelButtonTitle?: string;
    isConfirmButtonDisabled$?: Subject<boolean>;
    isCloseButtonHidden?: boolean;
    isCancelButtonHidden?: boolean;
    dialogSize?: DialogSizes;
    contentText?: string;
    contentComponent?: Type<any>;
    contentTemplate?: TemplateRef<any>;
    actionsTemplate?: TemplateRef<any>;
    descriptionTemplate?: TemplateRef<any>;
    headerIcon?: string;
    additionalActionButtons?: AdditionalDialogActionButton[];
    componentData?: any;
    dataOnConfirm$?: Subject<any>;
}

export interface AdditionalDialogActionButton {
    title: string;
    onClick: (args?: any) => void;
    class?: string;
}